// src/app/generated/model/modelCase.ts
/**
 * Covid19-case-follow-up
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ModelEnumIdValue } from './modelEnumIdValue';


export interface ModelCase { 
    id?: number;
    caseCode?: string;
    firstName: string;
    lastName: string;
    dob?: string;
    gender?: ModelCase.GenderEnum;
    occupation?: string;
    passportNumber: string;
    incidentContactPhone1: string;
    incidentContactPhone2?: string;
    countryOfResidence: string;
    countryOfOrigin?: string;
    nationality?: string;
    region: string;
    subcityOrZone?: string;
    woreda?: string;
    kebele?: string;
    houseNo?: string;
    phoneNo?: string;
    latitude?: string;
    longitude?: string;
    presumptiveResult?: ModelEnumIdValue;
    confirmedResult?: ModelEnumIdValue;
    identifiedBy: ModelEnumIdValue;
    parentCaseCode?: string;
    recentTravelTo?: string;
    admittedToFacility?: ModelEnumIdValue;
    status?: ModelEnumIdValue;
    modifiedBy: string;
    reportingDate?: Date;
    modifiedDate?: Date;
}
export namespace ModelCase {
    export type GenderEnum = 'F' | 'M';
    export const GenderEnum = {
        F: 'F' as GenderEnum,
        M: 'M' as GenderEnum
    };
}